Improve e2e test development ergonomics  (#155)
---------
