+++ test.env.example
@@ -1,0 +1,7 @@
+# Authentication details for running e2e tests 
+host=""
+http_path=""
+access_token=""
+
+# Only required to run the PySQLStagingIngestionTestSuite
+staging_ingestion_user=""
